LayoutTests/fast/dom/element-traversal-on-character-data-expected.txt

This test checks the implementation of the element traversal APIs on CharacterData as part of the NonDocumentTypeChildNode interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstTextNode.previousElementSibling is null
PASS firstTextNode.nextElementSibling is element
PASS firstTextNode.nextElementSibling.nextElementSibling is null
PASS lastTextNode.previousElementSibling is element
PASS lastTextNode.previousElementSibling.previousElementSibling is null
PASS lastTextNode.nextElementSibling is null
PASS successfullyParsed is true

TEST COMPLETE
